feat: move signer back
